requirements.txt
pyme3
PyGobject
xdo
